requirements.txt
flask==3.0.2
psycopg2-binary==2.9.8
python-dotenv==1.0.0